refactor(test): consolidate git-worktree-provider tests
Split .test.ts (93 tests) into focused files:
- .test.ts: 17 fault-injection tests using vi.fn() overrides
- .integration.test.ts: 77 behavioral tests (14 existing + 63 moved)

Dropped 13 tests: 3 Path utility tests (belong in path.test.ts),
4 internal duplicates, 6 cross-file duplicates (kept integration versions).
